requirements.txt
pyyaml
boto
pymongo
